Public/Private Access Control added

+++ serious_squad_app/views.py
@@ -23,6 +23,7 @@ import string
 import datetime,time
 from django.utils import timezone
 import threading
+import base64
 
 def superuser_only(function):
     def _inner(request, *args, **kwargs):
@@ -147,16 +148,15 @@ def clear_notification(request):
 def upload_data(request):
     date=datetime.datetime.now() + datetime.timedelta(days=1)
     date = date.strftime("%Y-%m-%dT%H:%M")
-    print(date)
     if request.method == 'POST':
         form = DataForm(request.POST, request.FILES,)
         if form.is_valid():
             fs=form.save(commit=False)
             fs.user_id = request.user.id
-            fs.key = generate_key()
+            fs.key = base64_encode(generate_key())
             fs.filename = fs.data.name
             fs.save()
-            encrypt_data(fs.data,fs.key)
+            encrypt_data(fs.data,base64_decode(fs.key))
             messages.info(request, 'Successfully Uploaded !')
             return redirect('upload_data')
     else:
@@ -181,20 +181,35 @@ def delete_data(request, id):
 def download_data(request, path ,data):
     file_path = MEDIA_ROOT+'/'+ path +'/' + data
     data_obj = Data.objects.get(data=path +'/' + data)
-    subprocess.run(f'openssl enc -d -aes-256-cbc -in '+'"'+file_path+'.enc'+'"'+' -out '+'"'+file_path+'"'+' -k '+data_obj.key+' -pbkdf2')
-    file_wrapper = FileWrapper(open(file_path,'rb'))
-    file_mimetype = mimetypes.guess_type(file_path)
-    response = HttpResponse(file_wrapper, content_type=file_mimetype )
-    response['X-Sendfile'] = file_path
-    response['Content-Length'] = os.stat(file_path).st_size
-    response['Content-Disposition'] = 'attachment; filename=%s' % smart_text(data)
-    os.remove(file_path)
-    return response
+    if request.user.id == data_obj.user_id or data_obj.universal == "Yes" :
+        subprocess.run(f'openssl enc -d -aes-256-cbc -in '+'"'+file_path+'.enc'+'"'+' -out '+'"'+file_path+'"'+' -k '+base64_decode(data_obj.key)+' -pbkdf2')
+        file_wrapper = FileWrapper(open(file_path,'rb'))
+        file_mimetype = mimetypes.guess_type(file_path)
+        response = HttpResponse(file_wrapper, content_type=file_mimetype)
+        response['X-Sendfile'] = file_path
+        response['Content-Length'] = os.stat(file_path).st_size
+        response['Content-Disposition'] = 'attachment; filename=%s' % smart_text(data)
+        os.remove(file_path)
+        return response
+    return render(request,'403.html')
 
 def generate_key():
-    characters = string.ascii_letters + string.digits
-    password = ''.join(random.choice(characters) for i in range(8))
+    symbols = "@#$="
+    characters = string.ascii_letters + string.digits + symbols
+    password = ''.join(random.choice(characters) for i in range(32))
     return password
+
+def base64_encode(string):
+    string_bytes = string.encode("ascii")
+    base64_bytes = base64.b64encode(string_bytes)
+    base64_string = base64_bytes.decode("ascii")
+    return base64_string
+
+def base64_decode(base64_string):
+    base64_bytes = base64_string.encode("ascii")
+    string_bytes = base64.b64decode(base64_bytes)
+    string = string_bytes.decode("ascii")
+    return string
 
 def encrypt_data(data,password):
     data = data.path
